Retry auth finalization if it fails the first time
There's some known issue in the auth flow where trying to auth twice in the same app session seems to fail the second time, particularly if there's a short turnaround between the end of the last session and the beginning of the new auth.

It seems to be related to trying to re-use a network connection that is now closed.

I'm not sure what the correct fix is, but this workaround addresses it by doing a single-shot retry:

More info here:

https://www.dropboxforum.com/t5/Dropbox-API-Support-Feedback/Re-linking-using-Obj-C-SDK-fails-with-network-error/td-p/593634