Fix the umbrella header warning
The root issue was that the "include" directory contained symlinks to all of the public headers
that are needed for both iOS and Mac builds. However, we were only building for iOS. So the warning
was telling us that there were publicly-facing headers that didn't make it into the final compiled
binary.

(I only vaguely understand this stuff, so that description might be somewhat flawed.)

To fix it, I:
 * Removed the `.macOS` target from the `Package.swift` to clarify that we only wanted that target
 * Removed the symlinks to the unused headers from the `include` directory

While I was in there, I tried my best to add values to `Package.swift` that were previously implicit.

I tried compiling by just removing the symlinks, and not making the changes to `path`, `publicHeadersPath`,
and `headerSearchPath`, but it wouldn't compile without those changes. I wish I understood why, but I don't.
It would have been nice if I could have included all the public headers in `headerSearchPath` (such as by
including the `include` directory that was full of symlinks), but I couldn't figure out how to do that.
I couldn't find any verbose build errors or warnings anywhere in the Xcode output.

Oh well. C'est la vie. This works and addresses the warnings, and that's good enough for now.