Keychain keys should survive a backup/restore
Currently, Dropbox stores the refresh token in the "device-only" section of the keychain.

That means that if the user transfers from an old device to a new device, the new device will be missing its Dropbox auth token and will deleech.

However, if we instead store the key in such a way that we can survive a backup, then we can have a seamless transition to the new device.

The downsides here are:

* once the new device refreshes the token, the old device will lose its auth
* if the old device refreshes the token before the new one can, then the new device need an auth

However, I am willing to accept these downsides, because the status-quo is a worse user experience than the potential downsides. At least with this change, the happy-path will be more seamless for the user.